requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
flask==3.0.0
Flask-Cors==4.0.0
future==0.18.3
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
numpy==1.24.4
python-whois==0.8.0
requests==2.31.0
scikit-learn==1.3.1
scipy==1.10.1
soupsieve==2.5
threadpoolctl==3.2.0
urllib3==2.0.6
werkzeug==3.0.0
zipp==3.17.0
